Create repository and update data structure
In this update I created two classes for repositories for players and bakers. Updated a bit of the player class and changed the data structure to hash map in order to have key value pair ith the player or baker and the value the corresponding name. Also implemented the menu system